Refactor, depth working
