+++ src/SearchAlgorithms/InformedSearch/AStar.py
@@ -23,7 +23,7 @@ class AStar(Avara):
         temp_First_goal = self.first_Goal
         temp_Second_goal = self.second_Goal
         while len(stack) != 0:
-            stack.sort(key=lambda node: node.getHeuristic(self.first_Goal, self.second_Goal)) # I sort by Heuristic (cost + manhattan), to get always the one of the minor value possible
+            stack.sort(key=lambda node: node.getHeuristic(self.first_Goal, self.second_Goal)) # CHANGE HERE COMPARED TO SUPER -> I sort by Heuristic (cost + manhattan), to get always the one with minor value(f = g + h) possible
             currentNode = stack.pop(0)
             currentNode.analizeGoal(temp_First_goal, temp_Second_goal)
